Refactor Numeric to be compliant with Ruby.
