+++ opal/opal.rb
@@ -22,7 +22,7 @@ require 'corelib/array/inheritance'
 require 'corelib/hash'
 require 'corelib/string/inheritance'
 require 'corelib/string/encoding'
-require 'corelib/numeric'
+require 'corelib/number'
 require 'corelib/complex'
 require 'corelib/rational'
 require 'corelib/math'
